backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
insightface==0.7.3
onnxruntime==1.17.0
onnx==1.15.0
